requirements.txt
psycopg2-binary>=2.9.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
configparser
hvac>=1.1.0